First iteration of fixes
