+++ components/virtual-node/sandbox-notary-impl/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.common-library'
+}
+
+description 'Corda Sandbox Notary'
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+    compileOnly 'org.osgi:org.osgi.service.component.annotations'
+
+    api 'net.corda:corda-notary-plugin'
+    implementation project(':components:ledger:notary-worker-selection')
+    implementation project(':libs:sandbox')
+    implementation project(':libs:virtual-node:sandbox-group-context')
+    implementation project(':components:virtual-node:sandbox-notary')
+
+    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
+}
