+++ settings.gradle
@@ -187,6 +187,7 @@ include 'components:virtual-node:sandbox-group-context-service:test-impl-one'
 include 'components:virtual-node:sandbox-group-context-service:test-impl-two'
 include 'components:virtual-node:sandbox-json'
 include 'components:virtual-node:sandbox-notary'
+include 'components:virtual-node:sandbox-notary-impl'
 include 'components:virtual-node:virtual-node-info-read-service'
 include 'components:virtual-node:virtual-node-info-read-service-rpc-extensions'
 include 'components:virtual-node:virtual-node-management-sender'
